Improve indentation performace
In big files indentation performance was really bad, due to
no limit for ParseOuterOperators() function. Add reasonable limit for
this function by finding `(` on first column in the line and previous
blank or comment line.

It still somewhat slow because of overuse of synstack() function.
I have an idea how to find balancing parens without using this function.
But this work postpone for some time.