An improved way of preserving and replaying original mappings
Certain mappings, such as scrolling in floating windows, try to be
non-intrusive, and when they can't be applied they use nvim_feedkeys()
to replay the original mapping.
However, this simple approach does not always work. For example, I have
<C-n> remapped to gt which stops working and instead scrolls by a
single line even when there are no floating windows.

This change brings a more sophisticated mechanism of capturing and
replaying the original mapping. The implementation was inspired by
anuvyklack/keymap-amend.nvim