Workaround for `E1510` being thrown by Neovim
There is a bug in Neovim that makes it throw `E1510` instead of `E474`
if the parsed JSON string is very large and incomplete.  This patch
works around the issue by also catching `E1510` and handling it the same
way as `E474`.  This patch can safely be reverted once the issue is
fixed in Neovim.

See also https://github.com/neovim/neovim/issues/29886